refactor: delete event buffer service shell
