--- backend/web/services/event_buffer.py
@@ -1,3 +1,0 @@
-"""Compatibility shell for thread runtime event buffers."""
-
-from backend.thread_runtime.events.buffer import *  # noqa: F403
