+++ tests/Unit/backend/web/services/test_event_buffer_owner.py
@@ -12,12 +12,11 @@ def test_thread_runtime_namespace_exports_legacy_helpers() -> None:
     sandbox_owner = importlib.import_module("backend.thread_runtime.sandbox")
     reads_owner = importlib.import_module("backend.thread_runtime.events.reads")
     buffer_owner = importlib.import_module("backend.thread_runtime.events.buffer")
-    buffer_shell = importlib.import_module("backend.web.services.event_buffer")
 
     assert history_owner.build_thread_history_transport is history_shell.build_thread_history_transport
     assert projection_owner.canonical_owner_threads is projection_shell.canonical_owner_threads
     assert convergence_owner.inspect_owner_thread_runtime is not None
     assert sandbox_owner.resolve_thread_sandbox is not None
     assert reads_owner.build_run_event_read_transport is not None
-    assert buffer_owner.ThreadEventBuffer is buffer_shell.ThreadEventBuffer
-    assert buffer_owner.RunEventBuffer is buffer_shell.RunEventBuffer
+    assert buffer_owner.ThreadEventBuffer is not None
+    assert buffer_owner.RunEventBuffer is not None
